refactor(libslic3r): extract Preset config-key constants, fix casts, const-correct Preset/PhysicalPrinter
- Config.cpp: explicit static_cast for the iterator-difference ->
  ConfigOptionEnumDef::value_to_index/label_to_index narrowing.
- Config.hpp: factor the repeated "abs(default) < 1e9 && (abs(default) >
  1e-9 || == 0)" assert condition used by every ConfigOptionFloats/Ints/
  FloatsOrPercents constructor into one valid_config_option_numeric_default()
  helper (BP1001, same condition everywhere); flatten
  ConfigOptionVector::resize's nested if/else-if ladder into early returns
  (BP1015); replace the by-index/by-value duplicated modify loop in the
  apply-diff path with a single apply_value lambda used both ways.
- Preset.cpp/.hpp: extract the repeated config-key string literals
  ("nozzle_diameter", "compatible_prints", "printhost_*", ...) into
  file-local opt_key_* constants (BP1001); make Preset::save() and
  PhysicalPrinter::save() const; take get_short_name/get_preset_name/
  type_from_name by const& instead of by value (BP1025); fix
  get_short_name/get_preset_name using an `int` for std::string::find's
  result (find returns npos, a huge size_t, which was being compared as
  `> 0` and could misbehave on no-match) -- now uses size_t with an
  explicit npos check; add the missing PhysicalPrinter::print_host_options()
  definition for the existing declaration in Preset.hpp.

libslic3r builds clean (0 errors).

+++ src/libslic3r/Config.cpp
@@ -896,7 +896,7 @@ std::optional<int> ConfigOptionEnumDef::value_to_index(const std::string &value)
 {
     assert(this->is_valid_open_enum() || this->is_valid_closed_enum());
     auto it = std::find(m_values.begin(), m_values.end(), value);
-    return it == m_values.end() ? std::optional<int>{} : std::optional<int>{it - m_values.begin()};
+    return it == m_values.end() ? std::optional<int>{} : std::optional<int>{static_cast<int>(it - m_values.begin())};
 }
 
 // Look up an index of label of this combo box. Used for open enums.
@@ -905,7 +905,7 @@ std::optional<int> ConfigOptionEnumDef::label_to_index(const std::string &value)
     assert(is_valid_open_enum());
     const auto &ls = this->labels();
     auto        it = std::find(ls.begin(), ls.end(), value);
-    return it == ls.end() ? std::optional<int>{} : std::optional<int>{it - ls.begin()};
+    return it == ls.end() ? std::optional<int>{} : std::optional<int>{static_cast<int>(it - ls.begin())};
 }
 
 std::optional<std::reference_wrapper<const std::string>> ConfigOptionEnumDef::enum_to_value(int enum_val) const
